% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionCellNet.R
\name{AddClinicalData}
\alias{AddClinicalData}
\title{Add clinical data}
\usage{
AddClinicalData(ExpCellNetObj, ClinicalData)
}
\arguments{
\item{ExpCellNetObj}{An ExpressionCellNet object.}

\item{ClinicalData}{A data.frame of clinical variables (rows=samples).}
}
\value{
Updated ExpressionCellNet object.
}
\description{
Optionally attaches clinical variables to the ExpressionCellNet object.
}
\examples{
\dontrun{
obj <- AddClinicalData(obj, clinical_df)
}
}
