% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionCellNet.R
\name{AddMetaData}
\alias{AddMetaData}
\title{Add sample-level metadata}
\usage{
AddMetaData(ExpCellNetObj, MetaData)
}
\arguments{
\item{ExpCellNetObj}{An ExpressionCellNet object.}

\item{MetaData}{A data.frame of metadata (rows=samples).}
}
\value{
Updated ExpressionCellNet object.
}
\description{
Optionally attaches sample-level metadata to the ExpressionCellNet object.
}
\examples{
\dontrun{
obj <- AddMetaData(obj, meta_df)
}
}
