% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionCellNet.R
\name{AnalyzeNetwork}
\alias{AnalyzeNetwork}
\title{Analyze network topology and identify hub genes}
\usage{
AnalyzeNetwork(ExpCellNetObj)
}
\arguments{
\item{ExpCellNetObj}{An ExpressionCellNet object.}
}
\value{
Updated ExpressionCellNet object with "Network Analysis" and "Hub Genes".
}
\description{
Computes degree, closeness, betweenness, and eigenvector centrality and stores:
(i) Network Analysis table and (ii) Hub Genes table (top 10%).
}
\examples{
\dontrun{
obj <- AnalyzeNetwork(obj)
}
}
