% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionCellNet.R
\name{GeneEnrichment}
\alias{GeneEnrichment}
\title{Functional enrichment analysis of network genes}
\usage{
GeneEnrichment(ExpCellNetObj)
}
\arguments{
\item{ExpCellNetObj}{An ExpressionCellNet object.}
}
\value{
A named list of enrichment result data.frames (per database).
}
\description{
Runs enrichment analysis for genes in the constructed network using enrichR libraries.
}
\examples{
\dontrun{
enr <- GeneEnrichment(obj)
}
}
