% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionCellNet.R
\name{LoadGenerations}
\alias{LoadGenerations}
\title{Load generations table}
\usage{
LoadGenerations(ExpCellNetObj, GenerationsData)
}
\arguments{
\item{ExpCellNetObj}{An ExpressionCellNet object.}

\item{GenerationsData}{A data.frame mapping genes to their generation labels.}
}
\value{
Updated ExpressionCellNet object.
}
\description{
Stores an externally provided gene-to-generation mapping as the "Generations" component.
The table should include columns Gene and Generation.
}
\examples{
\dontrun{
gen <- data.frame(Gene=c("PYGO2","UBAP2L"), Generation=c("Generation : 0","Generation : 1"))
obj <- LoadGenerations(obj, gen)
}
}
