% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionCellNet.R
\name{LoadNetwork}
\alias{LoadNetwork}
\title{Load network edge table}
\usage{
LoadNetwork(ExpCellNetObj, NetworkData)
}
\arguments{
\item{ExpCellNetObj}{An ExpressionCellNet object.}

\item{NetworkData}{A data.frame representing network edges (two columns: gene1, gene2).}
}
\value{
Updated ExpressionCellNet object.
}
\description{
Stores an externally provided edge list as the "Network" component of the object.
The table should include at least two columns corresponding to interacting gene symbols.
}
\examples{
\dontrun{
net <- data.frame(gene1=c("PYGO2","BRCA1"), gene2=c("UBAP2L","RAD51"))
obj <- LoadNetwork(obj, net)
}

}
