% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionCellNet.R
\name{LoadTrainTestSplit}
\alias{LoadTrainTestSplit}
\title{Load a predefined train/test split}
\usage{
LoadTrainTestSplit(ExpCellNetObj, TrainTestSplitData)
}
\arguments{
\item{ExpCellNetObj}{An ExpressionCellNet object.}

\item{TrainTestSplitData}{A data.frame defining sample split into Train/Test.}
}
\value{
Updated ExpressionCellNet object.
}
\description{
Replaces (or creates) the internal "Train-Test Samples" table in the object using a user-supplied split.
The input data.frame must include at least: Sample_Name, Sample_Number, Group (Train/Test).
}
\examples{
\dontrun{
split_df <- data.frame(
  Sample_Name = colnames(obj[["Count Matrix"]]),
  Sample_Number = seq_len(ncol(obj[["Count Matrix"]])),
  Group = c(rep("Train", 8), rep("Test", 2))
)
obj <- LoadTrainTestSplit(obj, split_df)
}
}
