% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionCellNet.R
\name{RegenerateTrainTest}
\alias{RegenerateTrainTest}
\title{Regenerate train/test split}
\usage{
RegenerateTrainTest(ExpCellNetObj)
}
\arguments{
\item{ExpCellNetObj}{An ExpressionCellNet object.}
}
\value{
Updated ExpressionCellNet object with regenerated split.
}
\description{
Re-creates the internal random 80/20 train-test split without stratification.
}
\examples{
\dontrun{
obj <- RegenerateTrainTest(obj)
}
}
