% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionCellNet.R
\name{ShowNetwork}
\alias{ShowNetwork}
\title{Visualize the constructed network}
\usage{
ShowNetwork(ExpCellNetObj)
}
\arguments{
\item{ExpCellNetObj}{An ExpressionCellNet object containing a "Network" edge table.}
}
\value{
A plot (igraph/base) is produced as a side effect.
}
\description{
Plots the current network stored in the ExpressionCellNet object.
}
\examples{
\dontrun{
obj <- BuildNetwork(obj, Gene="PYGO2", Generation=3, CorrelationThreshold=c(0.7,0.7,0.7))
ShowNetwork(obj)
}

}
